fix bug

+++ Android-Zjyiot/app/src/main/java/com/qyy/app/lipstick/model/response/home/GoodsBean.java
@@ -69,8 +69,8 @@ public  class GoodsBean {
         this.brand = brand;
     }
 
-    public float getCounterPrice() {
-        return counterPrice;
+    public int getCounterPrice() {
+        return (int) counterPrice;
     }
 
     public void setCounterPrice(float counterPrice) {
